Refactoring Equalizer Test
