+++ app/src/main/java/com/ece493/group5/adjustableaudio/ui/media_player/MediaPlayerFragment.java
@@ -10,6 +10,7 @@ import android.media.session.PlaybackState;
 import android.net.Uri;
 import android.os.Bundle;
 import android.provider.MediaStore;
+import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
